using the cholesky symmetry in VHS and cleaning up propagation/generic.py
