+++ ipie/legacy/propagation/tests/test_generic_phmsd.py
@@ -4,7 +4,7 @@ import pytest
 from ipie.systems.generic import Generic
 from ipie.legacy.hamiltonians.generic import Generic as HamGeneric
 from ipie.legacy.trial_wavefunction.multi_slater import MultiSlater
-from ipie.propagation.generic import GenericContinuous
+from ipie.legacy.propagation.generic import GenericContinuous
 from ipie.legacy.propagation.continuous import Continuous
 from ipie.utils.misc import dotdict
 from ipie.utils.testing import (
