+++ ipie/propagation/generic.py
@@ -1,13 +1,8 @@
-import cmath
-import math
+import time
 import numpy
 import scipy.linalg
-import sys
-import time
-from ipie.utils.linalg import exponentiate_matrix
-from ipie.legacy.walkers.single_det import SingleDetWalker
-from ipie.utils.linalg import reortho
 from ipie.utils.misc import is_cupy
+from ipie.utils.pack import unpack_VHS_batch
 
 class GenericContinuous(object):
     """Propagator for generic many-electron Hamiltonian.
@@ -31,14 +26,15 @@ class GenericContinuous(object):
     """
 
     def __init__(self, system, hamiltonian, trial, qmc, options={}, verbose=False):
-        self.optimised = options.get('optimised', True)
+        
+        assert(qmc.batched)
+
         # Derived Attributes
         self.dt = qmc.dt
         self.sqrt_dt = qmc.dt**0.5
         self.isqrt_dt = 1j*self.sqrt_dt
         start = time.time()
         if trial.ndets > 1:
-            self.optimised = False
             self.mf_shift = (
                     self.construct_mean_field_shift_multi_det(hamiltonian, trial)
                     )
@@ -54,30 +50,13 @@ class GenericContinuous(object):
         # Constant core contribution modified by mean field shift.
         self.mf_core = hamiltonian.ecore + 0.5*numpy.dot(self.mf_shift, self.mf_shift)
         self.nstblz = qmc.nstblz
-        if (qmc.batched):
-            self.nwalkers = qmc.nwalkers
-            self.vbias_batch = numpy.zeros((qmc.nwalkers, hamiltonian.nfields), dtype=numpy.complex128)
-        else:
-            self.vbias = numpy.zeros(hamiltonian.nfields, dtype=numpy.complex128)
-        if self.optimised:
-            if (qmc.batched):
-                self.construct_force_bias = None
-                self.construct_VHS_batch = self.construct_VHS_batch
-            else:
-                self.construct_force_bias = self.construct_force_bias_fast
-                self.construct_VHS = self.construct_VHS_fast
-        else:
-            if trial.ndets > 1:
-                if (qmc.batched):
-                    self.construct_force_bias = None
-                else:
-                    self.construct_force_bias = self.construct_force_bias_multi_det
-            else:
-                assert(qmc.batched == False or qmc.batched == None)
-                self.construct_force_bias = self.construct_force_bias_slow
-            self.construct_VHS = self.construct_VHS_slow
+
+        self.nwalkers = qmc.nwalkers
+        self.vbias_batch = numpy.zeros((qmc.nwalkers, hamiltonian.nfields), dtype=numpy.complex128)
+
         self.ebound = (2.0/self.dt)**0.5
         self.mean_local_energy = 0
+
         if verbose:
             print("# Finished setting up propagation.GenericContinuous.")
 
@@ -135,85 +114,6 @@ class GenericContinuous(object):
         self.BH1 = numpy.array([scipy.linalg.expm(-0.5*dt*H1[0]),
                                 scipy.linalg.expm(-0.5*dt*H1[1])])
 
-    def construct_force_bias_slow(self, hamiltonian, walker, trial):
-        """Compute optimal force bias.
-
-        Uses explicit expression.
-
-        Parameters
-        ----------
-        G: :class:`numpy.ndarray`
-            Walker's Green's function.
-
-        Returns
-        -------
-        xbar : :class:`numpy.ndarray`
-            Force bias.
-        """
-        # vbias = numpy.einsum('lpq,pq->l', hamiltonian.chol_vecs, walker.G[0])
-        # vbias += numpy.einsum('lpq,pq->l', hamiltonian.chol_vecs, walker.G[1])
-        vbias = numpy.dot(hamiltonian.chol_vecs.T, walker.G[0].ravel())
-        vbias += numpy.dot(hamiltonian.chol_vecs.T, walker.G[1].ravel())
-        return - self.sqrt_dt * (1j*vbias-self.mf_shift)
-
-    def construct_force_bias_fast(self, hamiltonian, walker, trial):
-        """Compute optimal force bias.
-
-        Uses rotated Green's function.
-
-        Parameters
-        ----------
-        Ghalf : :class:`numpy.ndarray`
-            Half-rotated walker's Green's function.
-
-        Returns
-        -------
-        xbar : :class:`numpy.ndarray`
-            Force bias.
-        """
-        G = walker.Ghalf
-        if hamiltonian.sparse:
-            self.vbias = trial.rot_chol(spin=0) * G[0].ravel()
-            self.vbias += trial.rot_chol(spin=1) * G[1].ravel()
-        else:
-            self.vbias = numpy.dot(trial.rot_chol(spin=0), G[0].ravel())
-            self.vbias += numpy.dot(trial.rot_chol(spin=1), G[1].ravel())
-        return - self.sqrt_dt * (1j*self.vbias-self.mf_shift)
-
-    def construct_force_bias_multi_det(self, hamiltonian, walker, trial):
-        # if (trial.G != None):
-        vbias = numpy.array([walker.contract_one_body(Vpq, trial)
-                             for Vpq in hamiltonian.chol_vecs.T])
-        return - self.sqrt_dt * (1j*vbias-self.mf_shift)
-        # else:
-        #     return self.construct_force_bias_slow(hamiltonian, walker, trial)
-
-    def construct_VHS_slow(self, hamiltonian, shifted):
-        # VHS_{ik} = \sum_{n} v_{(ik),n} (x-xbar)_n
-        nb = hamiltonian.nbasis
-        return self.isqrt_dt * numpy.einsum("mx,x->m", hamiltonian.chol_vecs, shifted).reshape(nb,nb)
-
-    def construct_VHS_fast(self, hamiltonian, xshifted):
-        """Construct the one body potential from the HS transformation
-        Parameters
-        ----------
-        hamiltonian :
-            hamiltonian class
-        xshifted : numpy array
-            shifited auxiliary field
-        Returns
-        -------
-        VHS : numpy array
-            the HS potential
-        """
-        if numpy.isrealobj(hamiltonian.chol_vecs):
-            VHS = numpy.einsum("mx,x->m", hamiltonian.chol_vecs, xshifted.real) + 1.j * numpy.einsum("mx,x->m",hamiltonian.chol_vecs, xshifted.imag)
-        else:
-            # VHS = hamiltonian.chol_vecs.dot(xshifted)
-            VHS = numpy.einsum("mx,x->m", hamiltonian.chol_vecs, xshifted)
-        VHS = VHS.reshape(hamiltonian.nbasis, hamiltonian.nbasis)
-        return  self.isqrt_dt * VHS
-
     def construct_VHS_batch(self, hamiltonian, xshifted):
         """Construct the one body potential from the HS transformation
         Parameters
@@ -234,139 +134,22 @@ class GenericContinuous(object):
         else:
             isrealobj = numpy.isrealobj
 
-        if isrealobj(hamiltonian.chol_vecs):
-            VHS = hamiltonian.chol_vecs.dot(xshifted.real) + 1.j * hamiltonian.chol_vecs.dot(xshifted.imag)
+        if (hamiltonian.symmetry):
+            if isrealobj(hamiltonian.chol_vecs):
+                VHS_packed = hamiltonian.chol_packed.dot(xshifted.real) + 1.j * hamiltonian.chol_packed.dot(xshifted.imag)
+            else:
+                VHS_packed = hamiltonian.chol_packed.dot(xshifted)
+            # (nb, nb, nw) -> (nw, nb, nb)
+            VHS_packed = VHS_packed.T.reshape(self.nwalkers, hamiltonian.chol_packed.shape[0]).copy()
+            VHS = numpy.zeros((self.nwalkers, hamiltonian.nbasis, hamiltonian.nbasis), dtype = VHS_packed.dtype)
+            # for iw in range(self.nwalkers):
+            unpack_VHS_batch(hamiltonian.sym_idx[0], hamiltonian.sym_idx[1], VHS_packed, VHS)
         else:
-            VHS = hamiltonian.chol_vecs.dot(xshifted)
-        # (nb, nb, nw) -> (nw, nb, nb)
-        VHS = VHS.T.reshape(self.nwalkers, hamiltonian.nbasis, hamiltonian.nbasis).copy()
+            if isrealobj(hamiltonian.chol_vecs):
+                VHS = hamiltonian.chol_vecs.dot(xshifted.real) + 1.j * hamiltonian.chol_vecs.dot(xshifted.imag)
+            else:
+                VHS = hamiltonian.chol_vecs.dot(xshifted)
+            # (nb, nb, nw) -> (nw, nb, nb)
+            VHS = VHS.T.reshape(self.nwalkers, hamiltonian.nbasis, hamiltonian.nbasis).copy()
+
         return  self.isqrt_dt * VHS
-
-def construct_propagator_matrix_generic(hamiltonian, BT2, config, dt, conjt=False):
-    """Construct the full projector from a configuration of auxiliary fields.
-
-    For use with generic hamiltonian object.
-
-    Parameters
-    ----------
-    hamiltonian : class
-        hamiltonian class.
-    BT2 : :class:`numpy.ndarray`
-        One body propagator.
-    config : numpy array
-        Auxiliary field configuration.
-    conjt : bool
-        If true return Hermitian conjugate of matrix.
-
-    Returns
-    -------
-    B : :class:`numpy.ndarray`
-        Full propagator matrix.
-    """
-    nbsf = hamiltonian.nbasis
-    VHS = 1j*dt**0.5*hamiltonian.chol_vecs.dot(config).reshape(nbsf, nbsf)
-    EXP_VHS = exponentiate_matrix(VHS)
-    Bup = BT2[0].dot(EXP_VHS).dot(BT2[0])
-    Bdown = BT2[1].dot(EXP_VHS).dot(BT2[1])
-
-    if conjt:
-        return [Bup.conj().T, Bdown.conj().T]
-    else:
-        return [Bup, Bdown]
-
-
-# def back_propagate(system, psi, trial, nstblz, BT2, dt):
-    # r"""Perform back propagation for RHF/UHF style wavefunction.
-
-    # For use with generic system hamiltonian.
-
-    # Parameters
-    # ---------
-    # system : system object in general.
-        # Container for model input options.
-    # psi : :class:`pie.walkers.Walkers` object
-        # CPMC wavefunction.
-    # trial : :class:`pie.trial_wavefunction.X' object
-        # Trial wavefunction class.
-    # nstblz : int
-        # Number of steps between GS orthogonalisation.
-    # BT2 : :class:`numpy.ndarray`
-        # One body propagator.
-    # dt : float
-        # Timestep.
-
-    # Returns
-    # -------
-    # psi_bp : list of :class:`pie.walker.Walker` objects
-        # Back propagated list of walkers.
-    # """
-    # psi_bp = [SingleDetWalker({}, system, trial, index=w) for w in range(len(psi))]
-    # nup = system.nup
-    # for (iw, w) in enumerate(psi):
-        # # propagators should be applied in reverse order
-        # for (i, c) in enumerate(w.field_configs.get_block()[0][::-1]):
-            # # could make this system specific to reduce need for multiple
-            # # routines.
-            # B = construct_propagator_matrix_generic(system, BT2, c, dt, True)
-            # psi_bp[iw].phi[:,:nup] = B[0].dot(psi_bp[iw].phi[:,:nup])
-            # psi_bp[iw].phi[:,nup:] = B[1].dot(psi_bp[iw].phi[:,nup:])
-            # if i != 0 and i % nstblz == 0:
-                # psi_bp[iw].reortho(trial)
-    # return psi_bp
-
-def back_propagate_generic(phi, configs, system, hamiltonian, nstblz, BT2, dt, store=False):
-    r"""Perform back propagation for RHF/UHF style wavefunction.
-
-    For use with generic system hamiltonian.
-
-    Parameters
-    ---------
-    system : system object in general.
-        Container for model input options.
-    psi : :class:`pie.walkers.Walkers` object
-        CPMC wavefunction.
-    trial : :class:`pie.trial_wavefunction.X' object
-        Trial wavefunction class.
-    nstblz : int
-        Number of steps between GS orthogonalisation.
-    BT2 : :class:`numpy.ndarray`
-        One body propagator.
-    dt : float
-        Timestep.
-
-    Returns
-    -------
-    psi_bp : list of :class:`pie.walker.Walker` objects
-        Back propagated list of walkers.
-    """
-    nup = system.nup
-    psi_store = []
-    for (i, c) in enumerate(configs.get_block()[0][::-1]):
-        B = construct_propagator_matrix_generic(hamiltonian, BT2, c, dt, False)
-        phi[:,:nup] = numpy.dot(B[0].conj().T, phi[:,:nup])
-        phi[:,nup:] = numpy.dot(B[1].conj().T, phi[:,nup:])
-        if i != 0 and i % nstblz == 0:
-            (phi[:,:nup], R) = reortho(phi[:,:nup])
-            (phi[:,nup:], R) = reortho(phi[:,nup:])
-        if store:
-            psi_store.append(phi.copy())
-
-    return psi_store
-
-def back_propagate_generic_bmat(system, psi, trial, nstblz):
-    r"""Perform back propagation for RHF/UHF style wavefunction.
-    """
-    psi_bp = [SingleDetWalker({}, system, hamiltonian, trial, index=w) for w in range(len(psi))]
-    nup = system.nup
-    for (iw, w) in enumerate(psi):
-        # propagators should be applied in reverse order
-        for (i, B) in enumerate(w.stack.stack[::-1]):
-            # could make this system specific to reduce need for multiple
-            # routines.
-            psi_bp[iw].phi[:,:nup] = numpy.dot(B[0].conj().T,
-                                               psi_bp[iw].phi[:,:nup])
-            psi_bp[iw].phi[:,nup:] = numpy.dot(B[1].conj().T,
-                                               psi_bp[iw].phi[:,nup:])
-            if i != 0 and i % nstblz == 0:
-                psi_bp[iw].reortho(trial)
-    return psi_bp
