+++ ipie/utils/pack.pyx
@@ -23,8 +23,8 @@ def pack_cholesky(
 def unpack_VHS(
         long[:] idx_i,
         long[:] idx_j,
-        double[:] VHS_packed,
-        double[:,:] VHS):
+        double complex[:] VHS_packed,
+        double complex[:,:] VHS):
 
     cdef long i, j
     cdef long nbsf, nut
@@ -34,6 +34,25 @@ def unpack_VHS(
     for i in range(nut):
         VHS[idx_i[i],idx_j[i]] = VHS_packed[i]
         VHS[idx_j[i],idx_i[i]] = VHS_packed[i]
+
+    return
+
+def unpack_VHS_batch(
+        long[:] idx_i,
+        long[:] idx_j,
+        double complex[:,:] VHS_packed,
+        double complex[:,:,:] VHS):
+
+    cdef long i, iw
+    cdef long nbsf, nut, nwalkers
+    nwalkers = VHS.shape[0]
+    nbsf = VHS.shape[1]
+    nut = round(nbsf *(nbsf+1)/2)
+
+    for iw in range(nwalkers):
+        for i in range(nut):
+            VHS[iw, idx_i[i],idx_j[i]] = VHS_packed[iw,i]
+            VHS[iw, idx_j[i],idx_i[i]] = VHS_packed[iw,i]
 
     return
